+++ src/components/ViewSource/ViewSource.js
@@ -146,6 +146,10 @@ const ViewSource = ({
   ], [stripes, redirectToMARCEdit]);
 
   useEffect(() => {
+    if (isInstanceLoading) {
+      return;
+    }
+
     setIsMarcLoading(true);
 
     const { okapi: { tenant, token, locale } } = stripes;
@@ -169,7 +173,7 @@ const ViewSource = ({
       .finally(() => {
         setIsMarcLoading(false);
       });
-  }, []);
+  }, [isInstanceLoading]);
 
   const canExport = !isHoldingsRecord && stripes.hasPerm('ui-data-export.edit');