+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fix Tags accordion displaying in Holdings and Items until Tags setting is re-enabled. Fixes UIIN-3602.
 * Make item status in Item details pane header translatable. Fixes UIIN-3632.
+* `<ViewSource>` - don't load MARC record until Instance has finished loading. Fixes UIIN-3635.
 
 ## [14.0.0](https://github.com/folio-org/ui-inventory/tree/v14.0.0) (2026-04-17)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v13.0.13...v14.0.0)